Refactor migration doctests and run as tests
Refactor the migration doc code into a utility function, and make new
script to run doctests only.  Use that script to test migration doctests
as part of the standard all-platform test runs.
