+++ .github/workflows/test-linux.yaml
@@ -128,6 +128,10 @@ jobs:
         run: |
           $PYTEST --doctest-plus --showlocals --pyargs skimage ./tests
 
+      - name: Run migration doctests
+        run: |
+          python doc/tools/doctest_migration.py
+
       - name: Check benchmarks
         run: |
           asv check -v -E existing
