+++ .github/workflows/test-windows.yaml
@@ -94,6 +94,10 @@ jobs:
         run: |
           $PYTEST --doctest-plus --showlocals --pyargs skimage ./tests
 
+      - name: Run migration doctests
+        run: |
+          python doc/tools/doctest_migration.py
+
       - name: "Job has failed: reporting"
         if: ${{ failure() && github.ref == 'refs/heads/main' }}
         uses: JasonEtco/create-an-issue@1b14a70e4d8dc185e5cc76d3bec9eab20257b2c5 # v2.9.2
